Moves production types to FeedPresentationModule
