+++ FeedModules/Presentation Layer/FeedPresentationModule/FeedPresenter.swift
@@ -1,0 +1,80 @@
+//
+//  FeedPresenter.swift
+//  FeedCoreModule
+//
+//  Created by Omran Khoja on 3/9/22.
+//
+
+import Foundation
+import FeedCoreModule
+
+public protocol FeedLoadingView {
+    func display(_ viewModel: FeedLoadingViewModel)
+}
+
+public struct FeedLoadingViewModel {
+    public let isLoading: Bool
+}
+
+public protocol FeedLoadingErrorView {
+    func display(_ viewModel: FeedLoadingErrorViewModel)
+}
+
+public struct FeedLoadingErrorViewModel {
+    public let errorMessage: String?
+    
+    static var noError: FeedLoadingErrorViewModel {
+        return FeedLoadingErrorViewModel(errorMessage: nil)
+    }
+    
+    static func error(message: String) -> FeedLoadingErrorViewModel {
+        return FeedLoadingErrorViewModel(errorMessage: message)
+    }
+}
+
+public protocol FeedView {
+    func display(_ viewModel: FeedViewModel)
+}
+
+public struct FeedViewModel {
+    public let feed: [FeedImage]
+}
+
+public final class FeedPresenter {
+    
+    private let loadingView: FeedLoadingView
+    private let errorView: FeedLoadingErrorView
+    private let feedView: FeedView
+    
+    public static var title: String {
+        return NSLocalizedString("FEED_VIEW_TITLE",
+                                 tableName: "Feed",
+                                 bundle: Bundle(for: FeedPresenter.self),
+                                 comment: "Title for the feed view")
+    }
+
+    public init(feedView: FeedView, loadingView: FeedLoadingView, errorView: FeedLoadingErrorView) {
+        self.loadingView = loadingView
+        self.errorView = errorView
+        self.feedView = feedView
+    }
+    
+    public func didStartLoadingFeed() {
+        errorView.display(.noError)
+        loadingView.display(FeedLoadingViewModel(isLoading: true))
+    }
+    
+    public func didFinishLoadingFeed(feed: [FeedImage]) {
+        feedView.display(FeedViewModel(feed: feed))
+        loadingView.display(FeedLoadingViewModel(isLoading: false))
+    }
+    
+    public func didFinishLoadingFeed(with error: Error) {
+        errorView.display(FeedLoadingErrorViewModel.error(message: feedLoadError))
+        loadingView.display(FeedLoadingViewModel(isLoading: false))
+    }
+    
+    private var feedLoadError: String {
+        return NSLocalizedString("FEED_VIEW_CONNECTION_ERROR", tableName: "Feed", bundle: Bundle(for: FeedPresenter.self), comment: "Error message displayed when we can't load the image feed from server.")
+    }
+}
